# src/openapi/addressbook-api.yaml
openapi: 3.0.0
info:
  title: Addressbook API
  description: |
    **Maintainer**: Cyd88
  version: 1.0.0
  contact:
    name: victor
    email: tuczkovski@gmail.com
servers:
  - url: /
paths:
...
  /v1/users:
    get:
      tags: [ users-v1 ]
      operationId: findUsers
      summary: Get a users.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/users/register:
    post:
      tags: [ users-v1 ]
      operationId: createUser
      summary: Add a user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/users/login:
    get:
      tags: [ users-v1 ]
      operationId: loginUser
      summary: Login user.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/users/{username}:
    get:
      tags: [ users-v1 ]
      operationId: findUserByUsername
      summary: Get a user by username.
      parameters:
        - name: username
          in: path
          required: true
          description: Username
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [ users-v1 ]
      operationId: updateUser
      summary: Update a user.
      parameters:
        - name: username
          in: path
          required: true
          description: Username
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/users/{secret}:
    get:
      tags: [ users-v1 ]
      operationId: findUserBySecret
      summary: Get a user by secret.
      parameters:
        - name: secret
          in: path
          required: true
          description: Secret
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      required: [ id, code, timestamp ]
      properties:
        id:
          type: string
          format: uuid
          description: Unique error ID, used for logging purposes, UUID format
        code:
          type: string
          description: A string coding the error type. This is given to caller so he can translate them if required.
        timestamp:
          type: integer
        message:
          type: string
          description: A short localized string that describes the error.
...
    User:
      title: Group
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        password:
          type: string
        secret:
          type: string
...